+++ examples/CHESS/src/pipeline/annotated/candidate_generation.py
@@ -51,11 +51,13 @@ output_format_instructions = \
 Please only provide a valid SQL query in a single string. Do not include any additional information or explanations.
 """
 
+lm_config = cognify.LMConfig(model="gpt-4o-mini", kwargs={"temperature": 0.0})
 exec = cognify.Model(agent_name="candidate_generation",
             system_prompt=system_prompt, 
-             inputs=[cognify.Input(name=input) for input in inputs], 
-             output=cognify.cognify.OutputLabel(name=output_format, custom_output_format_instructions=output_format_instructions))
-raw_runnable_exec = cognify.cognify.as_runnable(exec) | RawSqlOutputParser()
+             input_variables=[cognify.Input(name=input) for input in inputs], 
+             output=cognify.OutputLabel(name=output_format, custom_output_format_instructions=output_format_instructions),
+             lm_config=lm_config)
+raw_runnable_exec = cognify.as_runnable(exec) | RawSqlOutputParser()
 
 @chain
 def runnable_exec(input: dict):
