+++ examples/CHESS/src/pipeline/annotated/table_selection.py
@@ -44,9 +44,10 @@ Please respond with a JSON object structured as follows:
 Only output a json as your response.
 """   
 
-
+lm_config = cognify.LMConfig(model="gpt-4o-mini", kwargs={"temperature": 0.0})
 exec = cognify.Model(agent_name="table_selection",
              system_prompt=system_prompt, 
-             inputs=[cognify.Input(name=input) for input in inputs], 
-             output=cognify.OutputLabel(name=output_format, custom_output_format_instructions=output_format_instructions))
+             input_variables=[cognify.Input(name=input) for input in inputs], 
+             output=cognify.OutputLabel(name=output_format, custom_output_format_instructions=output_format_instructions),
+             lm_config=lm_config)
 runnable_exec = cognify.as_runnable(exec) | TableSelectionOutputParser()
